refactor id pages to DetailPageLayout
