+++ pages/keycaps/[id].js
@@ -1,8 +1,7 @@
 import { request, gql } from "graphql-request";
 import { PageLayout } from "@/layouts/PageLayout";
-import styles from "@/styles/Keycap.module.css";
+import { DetailPageLayout } from "@/layouts/DetailPageLayout";
 
-const strapiUrl = process.env.NEXT_PUBLIC_PROD_URL;
 const endpoint = process.env.PROD_GRAPHQL_ENDPOINT;
 
 export async function getStaticPaths() {
@@ -54,81 +53,13 @@ export async function getStaticProps({ params }) {
   };
 }
 
-function Tag({ children }) {
-  return <li className={styles.ListItem}>{children}</li>;
-}
-
 export default function Keycap({ data }) {
   return (
     <PageLayout data={data}>
-      <div className={styles.Container}>
-        {data.keycap.hero_image && (
-          <img
-            className={styles.MainImage}
-            src={`${strapiUrl}${data.keycap.hero_image.url}`}
-            alt={data.keycap.hero_image.alternativeText}
-          />
-        )}
-        <h2 className="fs--9 fw--normal ta--center">{data.keycap.name}</h2>
-        <ul
-          className={`${styles.List} ta--center mb--7`}
-          style={{ width: "100%" }}
-        >
-          {data.keycap.links && (
-            <>
-              {data.keycap.links[0].vendor && (
-                <Tag>
-                  <a href={data.keycap.links[0].vendor}>Vendor</a>
-                </Tag>
-              )}
-              {data.keycap.links[0].interest_check && (
-                <Tag>
-                  <a href={data.keycap.links[0].interest_check}>
-                    Interest Check
-                  </a>
-                </Tag>
-              )}
-            </>
-          )}
-        </ul>
-
-        <details className={styles.YouTube}>
-          <summary>Sound Test Video</summary>
-          <div className={styles.EmbedWrapper}>
-            <iframe
-              width="100%"
-              height="auto"
-              src="https://www.youtube.com/embed/g-vBr9a36NE"
-              frameBorder="0"
-              allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-              allowFullScreen
-            ></iframe>
-          </div>
-        </details>
-
-        <p className={`${styles.Post}`}>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Nostrum
-          perspiciatis corporis nemo natus veniam. Harum doloremque corrupti
-          velit! Debitis sint odit corrupti vel cum esse animi saepe facilis,
-          dignissimos quam! Lorem ipsum dolor sit amet consectetur adipisicing
-          elit. Nostrum perspiciatis corporis nemo natus veniam. Harum
-          doloremque corrupti velit! Debitis sint odit corrupti vel cum esse
-          animi saepe facilis, dignissimos quam! Lorem ipsum dolor sit amet
-          consectetur adipisicing elit. Nostrum perspiciatis corporis nemo natus
-          veniam. Harum doloremque corrupti velit! Debitis sint odit corrupti
-          vel cum esse animi saepe facilis, dignissimos quam! Lorem ipsum dolor
-          sit amet consectetur adipisicing elit. Nostrum perspiciatis corporis
-          nemo natus veniam. Harum doloremque corrupti velit! Debitis sint odit
-          corrupti vel cum esse animi saepe facilis, dignissimos quam!
-        </p>
-        <div className={styles.Gallery}>
-          <img
-            className={styles.Image}
-            src="/keyboards/photos/ava-yellow.jpg"
-            alt="ava"
-          />
-        </div>
-      </div>
+      <DetailPageLayout
+        heroImage={data.keycap.hero_image}
+        name={data.keycap.name}
+      />
     </PageLayout>
   );
 }
